feat(chat): combine application dialogs into one wizard
