+++ apps/chat/src/components/Common/ApplicationWizard/view-props.ts
@@ -1,0 +1,9 @@
+import { CustomApplicationModel } from '@/src/types/applications';
+
+export interface ViewProps {
+  isOpen: boolean;
+  onClose: (result: boolean) => void;
+  isEdit?: boolean;
+  currentReference?: string;
+  selectedApplication?: CustomApplicationModel;
+}
