+++ apps/chat/src/components/Marketplace/TabRenderer.tsx
@@ -26,9 +26,8 @@ import {
 import { FilterTypes, MarketplaceTabs } from '@/src/constants/marketplace';
 
 import { PublishModal } from '@/src/components/Chat/Publish/PublishWizard';
-import { ApplicationDialog } from '@/src/components/Common/ApplicationDialog';
+import { ApplicationWizard } from '@/src/components/Common/ApplicationWizard/ApplicationWizard';
 import { ConfirmDialog } from '@/src/components/Common/ConfirmDialog';
-import { QuickAppDialog } from '@/src/components/Common/QuickAppDialog';
 import ApplicationDetails from '@/src/components/Marketplace/ApplicationDetails/ApplicationDetails';
 import { CardsList } from '@/src/components/Marketplace/CardsList';
 import { MarketplaceBanner } from '@/src/components/Marketplace/MarketplaceBanner';
@@ -261,24 +260,13 @@ export const TabRenderer = ({ isMobile }: TabRendererProps) => {
       />
 
       {/* MODALS */}
-      {!!(
-        applicationModel && applicationModel.type === ApplicationType.CUSTOM_APP
-      ) && (
-        <ApplicationDialog
+      {!!applicationModel && (
+        <ApplicationWizard
           isOpen={!!applicationModel}
+          onClose={handleCloseApplicationDialog}
           isEdit={applicationModel.action === ApplicationActionType.EDIT}
           currentReference={applicationModel.entity?.reference}
-          onClose={handleCloseApplicationDialog}
-        />
-      )}
-      {!!(
-        applicationModel && applicationModel.type === ApplicationType.QUICK_APP
-      ) && (
-        <QuickAppDialog
-          isOpen={!!applicationModel}
-          isEdit={applicationModel.action === ApplicationActionType.EDIT}
-          currentReference={applicationModel.entity?.reference}
-          onClose={handleCloseApplicationDialog}
+          type={applicationModel.type}
         />
       )}
       {!!deleteModel && (
